Aminul

+++ app/Http/Controllers/Backend/Roles/RolesController.php
@@ -5,11 +5,24 @@ namespace App\Http\Controllers\Backend\Roles;
 use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Spatie\Permission\Models\Role;
 use Spatie\Permission\Models\Permission;
 
 class RolesController extends Controller
 {
+    public $user;
+
+
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            $this->user = Auth::guard('admin')->user();
+            return $next($request);
+        });
+    }
+
+
     /**
      * Display a listing of the resource.
      *
@@ -17,6 +30,10 @@ class RolesController extends Controller
      */
     public function index()
     {
+        if (is_null($this->user) || !$this->user->can('role.view')) {
+            abort(403, 'Sorry !! You are Unauthorized to view any role !');
+        }
+
         $roles = Role::all();
         return view('backend.role&permission.index', compact('roles'));
     }
@@ -28,6 +45,10 @@ class RolesController extends Controller
      */
     public function create()
     {
+        if (is_null($this->user) || !$this->user->can('role.create')) {
+            abort(403, 'Sorry !! You are Unauthorized to create any role !');
+        }
+
         $all_permissions = Permission::all();
         $permission_groups = User::getpermissionGroups();
         // dd($permission_groups);
@@ -42,13 +63,17 @@ class RolesController extends Controller
      */
     public function store(Request $request)
     {
+        if (is_null($this->user) || !$this->user->can('role.create')) {
+            abort(403, 'Sorry !! You are Unauthorized to create any role !');
+        }
+
         $request->validate([
             'name' => 'required|max:100|unique:roles'
         ], [
             'name.required' => 'please give a role name'
         ]);
 
-        $role = Role::create(['name' => $request->name]);
+        $role = Role::create(['name' => $request->name, 'guard_name' => 'admin']);
 
         $permissions = $request->input('permissions');
         if (!empty($permissions)) {
@@ -76,7 +101,11 @@ class RolesController extends Controller
      */
     public function edit($id)
     {
-        $role = Role::findById($id);
+        if (is_null($this->user) || !$this->user->can('role.edit')) {
+            abort(403, 'Sorry !! You are Unauthorized to edit any role !');
+        }
+
+        $role = Role::findById($id, 'admin');
         $all_permissions = Permission::all();
         $permission_groups = User::getpermissionGroups();
         // dd($permission_groups);
@@ -92,6 +121,9 @@ class RolesController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if (is_null($this->user) || !$this->user->can('role.edit')) {
+            abort(403, 'Sorry !! You are Unauthorized to edit any role !');
+        }
 
         $request->validate([
             'name' => 'required|max:100|unique:roles,name,' . $id
@@ -100,9 +132,9 @@ class RolesController extends Controller
         ]);
 
         // $role = Role::create(['name' => $request->name]);
-        $role = Role::findById($id,);
+        $role = Role::findById($id,'admin');
         $permissions = $request->input('permissions');
-        
+
         if (!empty($permissions)) {
             $role->name = $request->name;
             $role->save();
@@ -120,7 +152,11 @@ class RolesController extends Controller
      */
     public function destroy($id)
     {
-        $role = Role::findById($id);
+        if (is_null($this->user) || !$this->user->can('role.delete')) {
+            abort(403, 'Sorry !! You are Unauthorized to delete any role !');
+        }
+        
+        $role = Role::findById($id,'admin');
         if (!is_null($role)) {
             $role->delete();
         }
